main-service/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
aiosqlite
apscheduler
pytz
pydantic
email-validator
pydantic-settings
Jinja2
aiohttp
passlib[bcrypt]
fastadmin
python-multipart
python-jose
wordcloud
openpyxl
httpx
plotly
dash
pandas
